+++ HISTORY.rst
@@ -1,6 +1,31 @@
 History
 =======
 
+Version 1.0.12
+--------------
+- Basic S3 tasks.
+
+Version 1.0.11
+--------------
+- Improves RDS support.
+
+Version 1.0.10
+--------------
+- Adds retry support when adding tags in Ec2LaunchInstance.
+
+Version 1.0.9
+-------------
+- Adds name protection to Ec2LaunchInstance (prevent duplicate names).
+
+Version 1.0.8
+-------------
+- Bugfix release for 1.0.7.
+
+Version 1.0.7
+-------------
+- Experimental support for RDS
+- Experimental support for groups of AWS instances in the awsenv module.
+
 Version 1.0.6
 -------------
 - Helper function to create a remote file from string.